ObjectNode is moved to binding package
