+++ pyviews/binding/expression.py
@@ -1,19 +1,99 @@
 """Expression binding"""
 
+import ast
 from functools import partial
-from typing import Any, Callable, List
+from re import compile as compile_regex
+from typing import Any, Callable, Dict, List, NamedTuple, Union
 
 from pyviews.binding.binder import BindingContext
 from pyviews.core.bindable import Bindable, InheritedDict
 from pyviews.core.binding import Binding, BindingCallback, BindingError
 from pyviews.core.error import PyViewsError, error_handling
-from pyviews.core.expression import ATTRIBUTE, ENTRY, INDEX, Expression, ObjectNode, execute
+from pyviews.core.expression import Expression, execute
+
+ROOT = 'root'
+ENTRY = 'entry'
+ATTRIBUTE = 'attribute'
+INDEX = 'index'
 
 _GET_VALUE = {
     ENTRY: lambda inst, key: inst.get(key),
     ATTRIBUTE: getattr,
     INDEX: lambda inst, key: inst[key]
 } # yapf: disable
+
+_AST_CLASSES = {ast.Name, ast.Attribute, ast.Subscript}
+
+
+class ObjectNode(NamedTuple):
+    """Root entry in expression"""
+    key: str
+    type: str
+    children: List['ObjectNode']
+
+
+_OBJECT_NODE_CACHE: Dict[str, ObjectNode] = {}
+
+
+def _level_key(ast_node: ast.AST):
+    level = 0
+    while hasattr(ast_node, 'value'):
+        level = level + 1
+        ast_node = ast_node.value
+    return level
+
+
+def _get_index_value(ast_node: ast.Subscript) -> Any:
+    ast_node = ast_node.slice.value if isinstance(ast_node.slice, ast.Index) else ast_node.slice
+    if isinstance(ast_node, ast.Constant):
+        return ast_node.value
+    if isinstance(ast_node, ast.Num):
+        return ast_node.n
+    return _get_attr_expression(ast_node)
+
+
+def _get_attr_expression(ast_node: Union[ast.Name, ast.Attribute, ast.Subscript]) -> 'Expression':
+    result = ''
+    while isinstance(ast_node, ast.Attribute):
+        result = f'.{ast_node.attr}{result}'
+        ast_node = ast_node.value
+    return Expression(f'{ast_node.id}{result}')
+
+
+def _get_object_tree(code: str) -> ObjectNode:
+    if code in _OBJECT_NODE_CACHE:
+        return _OBJECT_NODE_CACHE[code]
+    root = ObjectNode('root', ROOT, [])
+    ast_root = ast.parse(code)
+    nodes: Dict[ast.AST, ObjectNode] = {}
+    ast_nodes = sorted((item for item in ast.walk(ast_root) if item.__class__ in _AST_CLASSES), key = _level_key)
+    for ast_node in ast_nodes:
+        if isinstance(ast_node, ast.Name):
+            try:
+                node = next(n for n in root.children if n.key == ast_node.id)
+            except StopIteration:
+                node = ObjectNode(ast_node.id, ENTRY, [])
+                root.children.append(node)
+            nodes[ast_node] = node
+        if isinstance(ast_node, ast.Attribute):
+            ast_parent = ast_node.value
+            try:
+                node = next(n for n in nodes[ast_parent].children if n.key == ast_node.attr)
+            except StopIteration:
+                node = ObjectNode(ast_node.attr, ATTRIBUTE, [])
+                nodes[ast_parent].children.append(node)
+            nodes[ast_node] = node
+        if isinstance(ast_node, ast.Subscript):
+            ast_parent = ast_node.value
+            key = _get_index_value(ast_node)
+            try:
+                node = next(n for n in nodes[ast_parent].children if n.key == key)
+            except StopIteration:
+                node = ObjectNode(key, INDEX, [])
+                nodes[ast_parent].children.append(node)
+            nodes[ast_node] = node
+    _OBJECT_NODE_CACHE[code] = root
+    return root
 
 
 class ExpressionBinding(Binding):
@@ -28,7 +108,7 @@ class ExpressionBinding(Binding):
 
     def bind(self, execute_callback = True):
         self.destroy()
-        objects_tree = self._expression.get_object_tree()
+        objects_tree = _get_object_tree(self._expression.code)
         with error_handling(BindingError, self._add_error_info):
             self._create_dependencies(self._vars, objects_tree)
         if execute_callback:
@@ -82,3 +162,45 @@ def bind_setter_to_expression(context: BindingContext) -> Binding:
     binding = ExpressionBinding(callback, expr, context.node.node_globals)
     binding.bind()
     return binding
+
+
+PROPERTY_EXPRESSION_REGEX = compile_regex(r'([a-zA-Z_0-9]{1,}\.){0,}([a-zA-Z_0-9]{1,})')
+
+
+def get_expression_callback(expression: Expression, expr_vars: InheritedDict) -> BindingCallback:
+    """Returns callback that sets value to property expression"""
+    _validate_property_expression(expression.code)
+    object_tree = _get_object_tree(expression.code)
+    if object_tree.children[0].children:
+        return partial(_property_expression_callback, object_tree, expr_vars)
+    return partial(_on_global_value_update, expr_vars, expression.code)
+
+
+def _validate_property_expression(source_code: str):
+    if not PROPERTY_EXPRESSION_REGEX.fullmatch(source_code):
+        error = BindingError('Expression should be property expression')
+        error.add_info('Expression', source_code)
+        raise error
+
+
+def _property_expression_callback(_var_tree: ObjectNode, _vars: InheritedDict, value: Any):
+    (inst, prop) = _get_target(_var_tree, _vars)
+    setattr(inst, prop, value)
+
+
+def _get_target(var_tree: ObjectNode, expr_vars: InheritedDict):
+    entry = var_tree.children[0]
+    inst = expr_vars[entry.key]
+    next_key = entry.children[0].key
+    entry = entry.children[0]
+
+    while entry.children:
+        inst = getattr(inst, next_key)
+        next_key = entry.children[0].key
+        entry = entry.children[0]
+
+    return inst, next_key
+
+
+def _on_global_value_update(_vars: InheritedDict, key: str, value):
+    _vars[key] = value
